+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
-
+## Unreleased
+- [CM-825] Fixed SPM integration issue
 ## [0.1.2] 2022-01-20
 - Fixed the Stroyboard linking issue
 ## [0.1.1] - 2022-01-11