+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
-## [Unreleased]
+## [0.2.1] - 2022-03-31
 - [CM-848] Added localisation support for last message of the conversation in Conversationlist screen.
+- [CM-773] Fixed thumbnail blobkey issues for photos & videos
 ## [0.2.0] - 2022-03-22
 - [CM-842] Added Support for s3 service for upload & download
 - [CM-825] Fixed SPM integration issue